requirements.txt
gitpython
huggingface_hub
jupyter
langchain
langchain-openai
openai
python-dotenv
streamlit